Separate the represented social responsibility from the Session root (#615)
Tenth delivery of the refactor push. 91 methods move into 5 bounded submodules
under session/social: group, guild, duel, trade and contacts. Every file is
under the 1,000-line review budget.

session/mod.rs drops from 39,283 to 37,808 production lines. Across the ten
deliveries the root falls from 75,841 to 37,808 - half of it retired into
navigable owners.

Methods that mention a group or trade but belong to spell, item, quest, loot,
creature, gameobject, pet or aura responsibilities already moved with those
families and are not touched again.

Safety established by comparison: the impl surface is 1,057 methods before and
after, an identical multiset, and the ownership baseline reports 3,761
associated items unchanged.

Verified: cargo test -p wow-world --lib 3,822 passed and 0 failed,
session-ownership-check --syntax-only PASS, check_architecture.py check and
self-test PASS, fmt and diff hygiene clean. Ceiling tightened.
