+++ crates/wow-world/src/session/social/mod.rs
@@ -1,0 +1,11 @@
+//! Represented social responsibility, separated from the
+//! Session root under #615. Each submodule owns one complete operation
+//! group; the canonical owners keep authority over the state they touch.
+
+use super::*;
+
+mod contacts;
+mod duel;
+mod group;
+mod guild;
+mod trade;
